+++ hata/discord/auto_moderation/tests/test__MentionSpamTriggerMetadata__utility.py
@@ -1,0 +1,29 @@
+import vampytest
+
+from .. import MentionSpamTriggerMetadata
+
+
+def test__MentionSpamTriggerMetadata__copy_0():
+    """
+    Tests whether ``MentionSpamTriggerMetadata``'s `copy` method works as expected.
+    Case: no mention limit.
+    """
+    metadata = MentionSpamTriggerMetadata(None)
+    
+    copy = metadata.copy()
+    
+    vampytest.assert_eq(metadata, copy)
+    vampytest.assert_is_not(metadata, copy)
+
+
+def test__MentionSpamTriggerMetadata__copy_1():
+    """
+    Tests whether ``MentionSpamTriggerMetadata``'s `copy` method works as expected.
+    Case: `20`.
+    """
+    metadata = MentionSpamTriggerMetadata(20)
+    
+    copy = metadata.copy()
+    
+    vampytest.assert_eq(metadata, copy)
+    vampytest.assert_is_not(metadata, copy)